requirements.txt
streamlit
pandas
joblib
folium
streamlit-folium
numpy
fpdf2
geopy
plotly
xgboost
scikit-learn